move MoonChain interfaces to MRL file
