+++ packages/config/src/types/AssetRoute.ts
@@ -4,7 +4,6 @@ import type {
   ContractConfigBuilder,
   ExtrinsicConfigBuilder,
   FeeConfigBuilder,
-  MrlConfigBuilder,
 } from '@moonbeam-network/xcm-builder';
 import type {
   AnyChain,
@@ -44,24 +43,8 @@ export interface FeeConfig {
   extra?: number;
 }
 
-export interface MrlConfig {
-  isAutomaticPossible: boolean;
-  transfer: MrlConfigBuilder;
-  moonChain: MoonChainConfig;
-}
-
 export interface DestinationFeeConfig
   extends SetOptional<FeeConfig, 'balance'> {
-  amount: number | FeeConfigBuilder;
-}
-
-export interface MoonChainConfig {
-  asset: Asset;
-  balance: BalanceConfigBuilder;
-  fee: MoonChainFeeConfig;
-}
-
-export interface MoonChainFeeConfig extends FeeConfig {
   amount: number | FeeConfigBuilder;
 }
